backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
# Phase 2: Real-time features
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
